coordination: validate that a transition's event is active in the transition's source state
